Added security, better file organization and update function for users
