+++ knt-backend/knt/internal/knt_database/receipt.go
@@ -4,13 +4,18 @@ import (
 	"database/sql"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"time"
 )
 
 //This file handles everything having to do with purchases and receipts
 
 func MakeTransaction(userId int, purchase PurchaseRequest, db *sql.DB) (int, error) {
+	//Begins the transaction
+	//This is important because if ANY error were to occur we need to reset the database to its original state
+	transaction, err := db.Begin()
+	if err != nil {
+		return 0, err
+	}
 	// get user
 	user, err := GetUser(db, userId)
 	if err != nil {
@@ -30,27 +35,33 @@ func MakeTransaction(userId int, purchase PurchaseRequest, db *sql.DB) (int, err
 		return 0, errors.New("Insufficient balance on user")
 	}
 	//generate receipt
-	receiptId, err := generateReceipt(db, purchase.Data)
+	receiptId, err := generateReceipt(transaction, purchase.Data)
 	if err != nil {
 		return 0, err
 	}
 	//make a transaction
-	err = generateTransaction(db, userId, user.Balance, cost, user.Balance-cost, receiptId)
+	err = generateTransaction(transaction, userId, user.Balance, cost, user.Balance-cost, receiptId)
 	if err != nil {
 		return 0, err
 	}
 	//subtract balance
-	err = setBalance(db, userId, user.Balance-cost)
+	err = setBalance(transaction, userId, user.Balance-cost)
 	if err != nil {
 		return 0, err
 	}
 
 	//Add tax
-	err = addTaxTotals(purchase.Data, db, cost)
+	err = addTaxTotals(purchase.Data, db, transaction, cost)
 	if err != nil {
+		transaction.Rollback()
 		return 0, err
 	}
 
+	//Finilize the transaction
+	err = transaction.Commit()
+	if err != nil {
+		return 0, err
+	}
 	return cost, nil
 }
 
@@ -76,13 +87,13 @@ func getProductValue(entry PurchaseEntry, db *sql.DB) (int, error) {
 }
 
 // Generated the receipt and stores it in the database
-func generateReceipt(db *sql.DB, entries []PurchaseEntry) (int64, error) {
+func generateReceipt(transaction *sql.Tx, entries []PurchaseEntry) (int64, error) {
 	dataString, err := json.Marshal(entries)
 	if err != nil {
 		return 0, err
 	}
 
-	receiptId, err := commitTransaction(db, "INSERT INTO receipts (data, timestamp) VALUES (?, datetime())", dataString)
+	receiptId, err := addToTransaction(transaction, "INSERT INTO receipts (data, timestamp) VALUES (?, datetime())", dataString)
 	if err != nil {
 		return 0, err
 	}
@@ -91,55 +102,51 @@ func generateReceipt(db *sql.DB, entries []PurchaseEntry) (int64, error) {
 }
 
 // Generates the transaction and stores it in the database
-func generateTransaction(db *sql.DB, userId int, startingBal int, deltaBal int, finalBal int, receiptId int64) error {
-	_, err := commitTransaction(db,
+func generateTransaction(transaction *sql.Tx, userId int, startingBal int, deltaBal int, finalBal int, receiptId int64) error {
+	_, err := addToTransaction(transaction,
 		"INSERT INTO transactions (user_id, starting_balance, delta_balance, final_balance, receipt_id) VALUES (?, ?, ?, ?, ?)",
 		userId, startingBal, deltaBal, finalBal, receiptId)
 	return err
 }
 
 // Sets the users balance to a specified ammount
-func setBalance(db *sql.DB, userId int, balance int) error {
-	_, err := commitTransaction(db, "UPDATE user SET balance = ? WHERE id = ?", 
-	balance, userId)
+func setBalance(transaction *sql.Tx, userId int, balance int) error {
+	_, err := addToTransaction(transaction, "UPDATE user SET balance = ? WHERE id = ?",
+		balance, userId)
 	return err
 }
 
 // Function designed to calculate total ammounts for tax reasons
 // The taxes are stored in their own tables in the form of tax{year}
-func addTaxTotals(entries []PurchaseEntry, db *sql.DB, cost int) error {
+func addTaxTotals(entries []PurchaseEntry, db *sql.DB, transaction *sql.Tx, cost int) error {
 	year := time.Now().Year()
 
 	//Ensure yearly tables existance
-	err := ensureTaxTableExists(db, year)
+	err := ensureTaxTableExists(transaction, year)
 	if err != nil {
-		fmt.Println("LOl")
 		return err
 	}
 	//Go through all the entries and apply the operation on all of them
 	for _, entry := range entries {
 		//check if entry exists
-		tax, err := getFirstEntry[TaxEntry](queryBuilder(db, "SELECT * FROM tax where year = ? and product_id = ?", 
-		year, entry.ProductId))
+		tax, err := getFirstEntry[TaxEntry](queryBuilder(db, "SELECT * FROM tax where year = ? and product_id = ?",
+			year, entry.ProductId))
 		if err != nil {
-			fmt.Println("LOl2")
 			return err
 		}
 		if tax.Id == 0 {
 			//The row for this product doesnt already exist in the table so create a new row for it
-			_, err = commitTransaction(db, "INSERT INTO tax (product_id, amount, totalCost, year) VALUES (?, ?, ?, ?)", 
-			entry.ProductId, entry.Amount, cost, year)
+			_, err = addToTransaction(transaction, "INSERT INTO tax (product_id, amount, totalCost, year) VALUES (?, ?, ?, ?)",
+				entry.ProductId, entry.Amount, cost, year)
 			if err != nil {
-				fmt.Println("LOl3")
 				return err
 			}
 			continue
 		}
 		//The product exists update it
-		_, err = commitTransaction(db, "UPDATE tax SET amount = ?, totalCost = ? WHERE year = ? and id = ?",
-		 tax.Amount+entry.Amount, tax.TotalCost+cost, year, tax.Id)
+		_, err = addToTransaction(transaction, "UPDATE tax SET amount = ?, totalCost = ? WHERE year = ? and id = ?",
+			tax.Amount+entry.Amount, tax.TotalCost+cost, year, tax.Id)
 		if err != nil {
-			fmt.Println("LOl4")
 			return err
 		}
 	}
@@ -147,7 +154,7 @@ func addTaxTotals(entries []PurchaseEntry, db *sql.DB, cost int) error {
 	return nil
 }
 
-func ensureTaxTableExists(db *sql.DB, year int) error {
-	_, err := commitTransaction(db, "CREATE TABLE IF NOT EXISTS tax (id INTEGER PRIMARY KEY AUTOINCREMENT, product_id INT, amount INT, totalCost INT, year INT)")
+func ensureTaxTableExists(transaction *sql.Tx, year int) error {
+	_, err := addToTransaction(transaction, "CREATE TABLE IF NOT EXISTS tax (id INTEGER PRIMARY KEY AUTOINCREMENT, product_id INT, amount INT, totalCost INT, year INT)")
 	return err
 }
