+++ knt-backend/knt/internal/knt_router/router.go
@@ -37,7 +37,7 @@ func assignAdminRoutes(r chi.Router, db *sql.DB, configRoutes *viper.Viper) {
 		assignAdminMiddleware(r, db)
 		r.MethodFunc(http.MethodGet, configRoutes.GetString("getUsersAdmin"), getUsersAdmin(db))
 		r.MethodFunc(http.MethodPost, configRoutes.GetString("createNewUser"), createNewUser(db))
-		r.MethodFunc(http.MethodPut, configRoutes.GetString("updateUser"), notImplemented)
+		r.MethodFunc(http.MethodPut, configRoutes.GetString("updateUser"), updateUser(db))
 		r.MethodFunc(http.MethodPost, configRoutes.GetString("updateUserMoney"), notImplemented)
 		r.MethodFunc(http.MethodGet, configRoutes.GetString("getUserAdmin"), getAdminUser(db))
